restaurant login opens modal view and sends the restaurant details and saves to DB
